+++ src/enums/dateType.ts
@@ -1,0 +1,10 @@
+export enum DateType {
+    Second = "Second",
+    Minute = "Minute",
+    Hour = "Hour",
+    Day = "Day",
+    Week = "Week",
+    Month = "Month",
+    Quarter = "Quarter",
+    Year = "Year"
+}