+++ src/enums/durationUnit.ts
@@ -1,0 +1,6 @@
+export enum DurationUnit {
+    Second = "second",
+    Minute = "minute",
+    Hour = "hour",
+    Day = "day",
+}