+++ src/enums/labelForDate.ts
@@ -1,0 +1,4 @@
+export enum LabelForDate {
+    Now = "Now",
+    Today = "Today"
+}