+++ src/enums/milestoneShape.ts
@@ -1,0 +1,5 @@
+export enum MilestoneShape {
+    Rhombus = "Rhombus",
+    Circle = "Circle",
+    Square = "Square"
+}