+++ src/enums/resourceLabelPosition.ts
@@ -1,0 +1,5 @@
+export enum ResourceLabelPosition {
+    Top = "Top",
+    Right = "Right",
+    Inside = "Inside"
+}